feat(lib-integration-tests): add integration tests comparing the output of our algorithm and of gsat
The saturation process is using excessive amount of heap space
and the test will likely not succeed on GitHub Actions.

I observed (through logs and profiling) that the heap usage
peaks within the join process and not in the rewriting process.
I think this is because we are unnecessarily remembering join result
for all 320K join rules output by our algorithm (I will further investigate
if we can mitigate this)