feat: reduce heap usage
The amount of heap required to produce a rewriting to the existential test case
dropped from >5GiB to <300MiB. The issue was that every time we generate
a Datalog rule from subquery entailment instance, we were using fresh variables,
and that was rendering atom caching useless. Now that we prepare variable context
for each rule we generate, we are able to make a bunch of Datalog rules
refer to the same atom instance (through the global cache), drastically reducing
the heap allocation.