feat: optimize output rule evaluation by splitting the output into input-rule saturation and goal derivation
This change did not really affect simpleQuery_0_existentialJoinQuery testcase
because the Datalog saturation of simpleRule_0 is not recursive (so the saturation
completes in a single run anyway).
In more general settings where recursive rules are formed by GSat
this change should affect performance, and in any case the optimization
does not really harm.