fix: only include constants appearing exclusively in relevant subquery as subquery constants
The previous implementation actually violated the contract of
SubqueryEntailmentComputation, because the query constant embeddings
should be keeping track of which constants, appearing in the query
and not in the rule, are embedded in the test instance to validate the
query atom containing those constants.

This fixes the bug identified in 73e2d892.