refactor: flip the LHS and RHS of the "instance strength relation".
Consider a local instance I1 that is homomorphic to I2 (relative to a set L of local names), that is, a pair (I1, I2) of local instances such that there is a local-name homomorphism σ: I1 → I2 such that L ⊆ fixpoints(σ).

I initially thought that it makes sense to call I1 "stronger than" I2, since if we consider the local instances to be "assumptions" to fire a tentacle of a certain shape, then having a smaller, un-unified instance (in this case I1) is more powerful than having a larger, unified instance (I2).
This terminology is, however, confusing, since we will often have to talk about the strength of local instances (without regarding them as "assumptions"), and in such scenario, calling a smaller instance "stronger" is inappropriate. If we want to talk about local instances regarded as assumptions, we could always say something like "the weaker the instance is, the stronger it is as an assumption".